# backend/src/main/java/es/codeurjc/daw/museum/dto/MuseumObjectDTO.java
package es.codeurjc.daw.museum.dto;

import java.util.List;

public record MuseumObjectDTO(
    Long id,
    String objectName,
    String groupName,
    String technicalData,
    String description,
    String category,
    List <NoteDTO> notes,
    boolean isSeen
) {}